64bit Server.exe

+++ server/EQGameScanner.h
@@ -1,4 +1,4 @@
-/* 
+/*
  * Smart EQ Offset Finder - GPL Edition
  * Copyright 2007-2009, Carpathian <Carpathian01@gmail.com>
  *
@@ -36,7 +36,7 @@ public:
 	bool executableExists() const;
 	void setExe(TCHAR* str);
 	DWORD findEQPointerOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask);
-	DWORD findEQStructureOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask, const DWORD baseEQPointerAddress);
+	DWORD findEQStructureOffset(DWORD startAddress, std::size_t blockSize, const PBYTE byteMask, const PCHAR charMask, const QWORD baseEQPointerAddress);
 	bool ScanExecutable(HWND hDlg, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf, bool write_out = false);
 	void ScanSecondary(HWND hDlg, IniReaderInterface* ir_intf, NetworkServerInterface* net_intf);
